backend/requirements.txt
aiofiles==0.6.0
Authlib==0.14.3
fastapi==0.65.1
celery==5.0.0
colormap==1.0.3
easydev==0.11.1
redis==3.5.3
httpx==0.15.5
ipython==7.18.1
itsdangerous==1.1.0
Jinja2==2.11.2
psycopg2==2.8.6
pytest==6.1.0
requests==2.24.0
starlette==0.14.2
uvicorn==0.12.1
passlib==1.7.2
bcrypt==3.2.0
pyjwt==1.7.1
mongoengine==0.22.1
pymongo==3.11.2
sendgrid==6.4.7
watchdog==0.10.2
python-multipart==0.0.5
apscheduler==3.7.0
wikipedia==1.4.0
python-dateutil==2.8.1